requirements.txt
fastapi
uvicorn
python-dotenv
strands-agents
strands-agents-tools
ollama
